+++ thesis/src/distance/EuclideanDistance.java
@@ -2,7 +2,7 @@ package distance;
 
 public class EuclideanDistance extends BidimensionalDistance {
 
-	public EuclideanDistance(double x1, double y1, double x2, double y2) {
+	public EuclideanDistance(short x1, short y1, short x2, short y2) {
 		super(x1, y1, x2, y2);
 	}
